Added SDL2_ttf. Added TextComponent. Added simple ui for scores and lives. Cleaning glm

+++ SDLTest-master/SDLGameTest/game/EnemyComponentMgr.cpp
@@ -69,7 +69,7 @@ Entity EnemyComponentMgr::checkShot(const SDL_Rect& rect)
 
 void EnemyComponentMgr::allocate(size_t size)
 {
-	m_data.life = static_cast<size_t*>(MemoryPool::Get(size * sizeof(size_t)));
+	m_data.life = static_cast<int32_t*>(MemoryPool::Get(size * sizeof(int32_t)));
 	m_data.position = static_cast<vec3*>(MemoryPool::Get(size * sizeof(vec3)));
 	m_data.entity = static_cast<Entity*>(MemoryPool::Get(size * sizeof(Entity)));
